first pass; organizing process into methods and adding notes for further research.
The structure is being filled out thinking these woould be the common options for modulation; I don't need to understand what libraries are needed yet, though I have a clue.

The question is the processing, the runtime options instrumented and the needed data going where it likely should -- broken into composable pieces to produce a result.

Not going overboard with this, just basic structure.