More editorial; this appears to be better organized now, but there are opportunities for expansion of ideas AND typos to deal with. See TT & EM.
The formatting is coming along; I like the idea of the chapter header and overview being in a box.

The figures in the tables still need cites and attributions as do any links.

The idea today is to paint broad strokes and get the idea down; I have full confidence this ccan be achieved, the question is are my hunches right and is it now feasible to do so.